==> 02_bacdive_taxonomy.Rmd <==
---
title: "2H SIP Taxonomy"
author: "Tristan Caro"
date: "10/26/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load Packages
```{r}
library(tidyverse)
library(ggsci)
library(ggbiplot)
library(readxl)
library(ggrepel)
```

# Load Data
```{r}
# Generation time data from `qSIP_calculations.Rmd`
gen_calc_data_summarized <- read_xlsx(file.path("data", "gen_calc_data_summarized.xlsx"))

# BacDive data from `fatty_acids_from_BacDive.Rmd`
fatty_acids_with_tax <- read_xlsx(file.path("data", "fatty_acids_with_tax.xlsx"))

fatty_acids_wide_with_tax <- read_xlsx(file.path("data", "fatty_acids_wide_with_tax.xlsx"))
```

# Join dataframes

```{r}
fatty_acids_with_tax <- fatty_acids_with_tax %>% 
  mutate(
    FA = case_when(
      FA == "a15:0" ~ "a-15:0",
      FA == "i15:0" ~ "i-15:0",
      FA == "a17:0" ~ "a-17:0",
      FA == "i17:0" ~ "i-17:0",
      TRUE ~ FA
    ))

sixteen_ones <- c("16:1 trans 9", "16:1 cis 9")
eighteen_ones <- c("18:1 cis 9", "18:1 trans 9", "18:1")


gen_calc_taxa <- gen_calc_data_summarized %>% 
  select(terrain, inc_time_d, compound, at2H, at2H_mn_zero, rel_area_FID, u_d, gen_d, chain_fctr, chain_num) %>% 
  mutate(
    FA = case_when(
      compound %in% sixteen_ones ~ "16:1",
      compound == "16:0" ~ "16:0",
      compound == "18:0" ~ "18:0",
      compound %in% eighteen_ones ~ "18:1",
      compound == "18:2 trans 9, 12" ~ "18:2",
      compound == "17:0 cyclo" ~ "cyc 17:0",
      TRUE ~ compound
  ))

## Making the choice to select 3 day turnover rates, averaged across environments
gen_calc_taxa_3d_agg <- gen_calc_taxa %>% 
  filter(inc_time_d == 3) %>% 
  group_by(FA) %>% 
  summarize_if(is.numeric, mean, na.rm = TRUE)

FA_w_turnover <- left_join(fatty_acids_with_tax, gen_calc_taxa_3d_agg, by = "FA")
```


# Export

```{r}
fatty_acids_wide_with_tax %>% saveRDS(file = "cache/bacdive_FA_wide.RDS")
write_csv(fatty_acids_wide_with_tax, file = "data_output/bacdive_FAs.csv")

fatty_acids_with_tax %>% saveRDS(file = "cache/bacdive_FA.RDS")

```
