==> 03_data_viz.Rmd <==
---
title: "03 Data Vizualization"
author: "Tristan Caro"
date: "`r Sys.Date()`"
output:
    html_document:
    code_folding: show
    df_print: paged
    number_sections: yes
    toc: yes
    toc_depth: 3
    toc_float: yes
editor_options:
  chunk_output_type: console
---

# Clear the environment

```{r}
rm(list=ls())
```

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load libraries

```{r, echo = FALSE, warning = FALSE, message=FALSE}
library(tidyverse)    # CRAN v1.3.1    
library(forcats)      # CRAN v0.5.1      
#library(isoreader)    # CRAN v1.3.0    
#library(isoprocessor) # [github::isoverse/isoprocessor] v0.6.7 
#library(isotopia)     # [github::isoverse/isotopia] v0.5.8     
library(readxl)       # CRAN v1.3.1       
library(ggsci)        # CRAN v2.9        
library(ggborderline) # [github::wurli/ggborderline] v0.1.0 
library(ggrepel)      # CRAN v0.9.1      
library(latex2exp)    # CRAN v0.5.0    
library(ggdist)       # CRAN v3.0.0       
library(ggsignif)     # CRAN v0.6.2     
library(ggridges)     # CRAN v0.5.3     
library(cowplot)      # CRAN v1.1.1      
library(scales)       # CRAN v1.1.1
library(ggforce)      # Accelerating 'ggplot2', CRAN v0.3.3

```

## Load sourced functions

```{r}
# Sourced Functions
source(file.path("libs", "visualization.R"))
source(file.path("libs", "chromeleon.R"))
source(file.path("libs", "error_prop.R"))
source(file.path("libs", "calculate_turnover.R"))
source(file.path("libs", "turnover_to_gen.R"))
source(file.path("libs", "rename_FAs.R"))
source(file.path("libs", "d2H_to_2F.R"))
`%nin%` = Negate(`%in%`) # "Not In" function
var_to_str = function(v) {return(deparse(substitute(v)))} # Convert variable name to string
```




## Read RDS from cache

```{r workspace images}
# Turnover rates
turnover <- readRDS("cache/turnover_rates.RDS")

# Turnover rates averaged across experimental replicates
turnover_summarized <- readRDS("cache/turnover_summarized.RDS")

# Turnover rates averaged across 3d, 7d, experimental replicates
turnover_summarized_avg <- readRDS("cache/turnover_summarized_avg.RDS")

# Abundance-weighted turnover rates
turnover_weighted <- readRDS("cache/turnover_weighted.RDS")

# Abundance-weighted turnover rates averaged between 3d and 7d incubation time points
turnover_weighted_avg <- readRDS("cache/turnover_weighted_avg")

# LH-SIP data
LH_SIP_data <- readRDS(file = "cache/LH_SIP_data.RDS")

# BacDive Data
bacdive_FA <- readRDS(file = "cache/bacdive_FA.RDS")
bacdive_FA_wide <- readRDS(file = "cache/bacdive_FA_wide.RDS")

# IRMS chromatogram data
sample_files <- readRDS(file = "cache/sample_files.RDS")

# IRMS Error Estimation
sF_0 <- readRDS(file = "cache/sF_0.RDS")

```

### Read lit analysis data

```{r lit analysis data}
lit_turnover <- readxl::read_excel(file.path("data", "literature_turnover_rates.xlsx"))

turnover_means <- turnover_weighted %>% 
  filter(inc_time_d == 7) %>%
  select(terrain, u_d_mc_weighted) %>% 
  rename(u_d_mn = u_d_mc_weighted) %>% 
  mutate(comment = terrain) %>% 
  mutate(n = case_when(
    terrain == "Niwot Ridge Tundra" ~ 3,
    terrain == "Gordon Gulch Conifer Forest" ~ 1,
    terrain == "Marshall Mesa Grassland" ~ 2
  )) %>% 
  mutate(method = "LH-SIP",
         method_type = "LH-SIP")

lit_turnover_w_ours <- lit_turnover %>% 
  bind_rows(turnover_means) %>% 
  # Convert to generation time:
  mutate(gen_d = turnover_to_gen(u_d_mn)) %>% 
  filter(method_type != "Microscopy")
```

### Load d2H assimilation data

```{r}
d2H_assim <- read_xlsx("data/d2H_data_isotope_measurements.xlsx", sheet = 'lipids')
```


### Read detection limit data

```{r}
dl_df <- readxl::read_excel("data/detection_limit_calcs.xlsx")
```


## Set color swatches

```{r color swatches}
tundra_color <- "#4ebad5"

grassland_color <- "#01a088"

conifer_color <- "#e64a35"

terrain_color_swatch <- c(conifer_color, grassland_color, tundra_color)
terrain_fill_swatch <- c(conifer_color, grassland_color, tundra_color)
```


# Plots



## F1: dF and growth rate Plot

### 1A: dF plot

Using memory-corrected values

```{r}
p_dF <- turnover_summarized %>% 
  filter(
    compound %nin% c("H2", "20:0", "22:0", "24:0", "a-15:0")) %>% 
  ungroup() %>% 
  select(terrain, inc_time_d, compound, dF_mc, calibrated_at2H_with_memory_correction_se) %>%
  left_join(
    turnover_weighted %>% 
      select(terrain, inc_time_d, dF_weighted_mc), 
    by = c("terrain", "inc_time_d")) %>% 
  group_by(
    terrain,
    inc_time_d
  ) %>% 
  pivot_longer(
    cols = c(dF_mc, dF_weighted_mc),
    values_to = "dF", 
    names_to = "type"
  ) %>% 
  mutate(compound = case_when(
    type == "dF_weighted_mc" ~ "Weighted Mean",
    TRUE ~ compound
  )) %>%
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>%

  ggplot(
    aes(
      x = inc_time_d,
      y = dF * 10000, # convert from at% to ppm
      group = compound,
      color = terrain,
      fill = terrain,
      alpha = type,
      size = type
    )
  ) +
  geom_point(alpha = 1) +
  geom_line() +
  facet_wrap(vars(terrain), scales = "free_y") +
  scale_color_manual(values = terrain_color_swatch) +
  scale_fill_manual(values = terrain_fill_swatch) +
  scale_alpha_discrete(range = c(1, 1)) +
  scale_size_discrete(range = c(0.25, 2)) +
  scale_x_continuous(breaks = c(0, 1,2,3,4,5,6,7)) +
  coord_cartesian(ylim = c(0, 2500)) +
  theme_classic() +
  labs(
    x = "Time elapsed (days)",
    y = TeX("$ \\Delta ^2H$ (ppm)")
  ) +
  theme(
    legend.position = "none",
    strip.background = element_blank(),
    strip.text = element_text(size = 14),
    plot.title = element_text(hjust = 0.5),
    axis.title.x = element_text(size = 10),
    panel.grid.major.y = element_line(linetype = "dotted"),
    panel.spacing = unit(1.75, "cm")
  )
p_dF


#plotly::ggplotly()
# 
# cowplot::save_plot(
#   filename = "fig_output/p_dF.pdf",
#   plot = p_dF,
#   base_width = 11,
#   base_height = 4
#   )
#   
```


### 1B: Growth rate plot

```{r turnover plot}
# Clean up weighted turnover data
turnover_weighted_cleaned <- turnover_weighted %>% 
  filter(inc_time_d == 7) %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>% 
  mutate(
    compound = "Weighted Mean",
    type = "Weighted"
  ) %>% 
  rename(
    "u_d_mc" = u_d_mc_weighted,
    "u_d" = u_d_weighted,
    "gen_d" = gen_d_weighted,
    "gen_d_mc" = gen_d_mc_weighted
  ) %>%
  select(terrain, compound, u_d_mc, u_d, gen_d, gen_d_mc, type)
  

# Define a function for turnover plots
plot_turnover <- function(df, terrain_choice, color_choice) {
  df %>% 
  # Data organization and cleanup
  ungroup() %>% 
  filter(
    inc_time_d == "7",
    compound %nin% c("H2", "20:0", "22:0", "24:0", "i-16:0")) %>%
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra")
    ) %>% 
  mutate(
    compound = str_replace(
      string = compound,
      pattern = "ω",
      replacement = "w")
  ) %>%
  select(terrain, compound, inc_time_d, u_d, u_d_mc, gen_d, gen_d_mc, su_d, su_d_mc) %>%
  mutate(type = "unweighted") %>% 
  bind_rows(
    turnover_weighted_cleaned) %>%
  group_by(terrain) %>% 
  
  # Choose terrain
  filter(terrain == terrain_choice) %>% 
  mutate(compound = forcats::fct_reorder(compound, u_d_mc)) %>% 
  
  # Plot it
  ggplot(
    aes(
      x = u_d_mc,
      y = compound,
      color = type,
      fill = type,
      label = round(gen_d_mc, 2)
      )
  ) +
    geom_point(size = 3) +
    geom_errorbar(aes(
      xmin = u_d_mc - (0.5 * su_d_mc),
      xmax = u_d_mc + (0.5 * su_d_mc)),
      width = 0,
      size = 1.1
      ) +
    # #Uncomment to show generation times
    # geom_text(
    #   nudge_x = 0.05
    # ) +
    ggrepel::geom_text_repel(
    data = function(df) df |> filter(compound == "Weighted Mean"),
    aes(
      label = paste(formatC(u_d, format = "e", digits = 1), 
                    " = \n", round(gen_d_mc, 0), "days")
      ),
    nudge_x = 0.1,
    nudge_y = -1,
    color = "red",
    size = 3,
    segment.alpha = 0.5, segment.curvature = -1e-20
    
  ) +
    coord_cartesian(xlim = c(0, 0.25)) +
    scale_fill_manual(values = c("black", color_choice)) +
    scale_color_manual(values = c("black", color_choice)) +
    theme_classic() +
    labs(
      x = TeX("Compound-specific growth rate (day$^{-1}$)"),
      title = terrain_choice
      ) +
    theme(
      axis.title.y = element_blank(),
      axis.text.x = element_text(angle = 45, hjust = 1),
      axis.ticks.y = element_blank(),
      strip.background = element_blank(),
      legend.position = "none",
      plot.title = element_text(hjust = 0.5),
      panel.grid.major.x = element_line(linetype = "dotted"),
      aspect.ratio = 1
    )

  }

p_tundra_turnover <- plot_turnover(
  turnover_summarized, 
  terrain_choice = "Tundra", 
  color_choice = tundra_color)

p_grassland_turnover <- plot_turnover(
  turnover_summarized,
  terrain_choice = "Grassland",
  color_choice = grassland_color
)

p_conifer_forest_turnover <- plot_turnover(
  turnover_summarized,
  terrain_choice = "Conifer Forest",
  color_choice = conifer_color
)

p_turnover <- cowplot::plot_grid(
  p_conifer_forest_turnover,
  p_grassland_turnover,
  p_tundra_turnover,
  nrow = 1
)

p_turnover

ggsave(
  plot = p_turnover,
  filename = "fig_output/p_turnover2.pdf",
  #device = cairo_pdf, # Need this for greek symbols to render correctly
  width = 11,
  height = 4
  )

```

### Cowplot

```{r}
p_dF_ud <- cowplot::plot_grid(
  p_dF, p_turnover,
  nrow = 2,
  labels = "AUTO"
)
p_dF_ud

ggsave(filename = "fig_output/p_dF_ud.pdf", 
       plot = p_dF_ud,
       device = cairo_pdf, # required for rendering unicode symbols
       width = 11,
       height = 8)

```


```{r}

turnover_summarized %>% 
  # Data organization and cleanup
  ungroup() %>% 
  filter(
    inc_time_d == "7",
    compound %nin% c("H2", "20:0", "22:0", "24:0", "i-16:0")) %>%
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>% 
  select(terrain, compound, inc_time_d, u_d, u_d_mc, gen_d, gen_d_mc, su_d, su_d_mc) %>%
  bind_rows(
    turnover_weighted_cleaned) %>%
  group_by(terrain) %>% 
  
  # Choose terrain
  filter(terrain == "Tundra") %>% 
  mutate(compound = forcats::fct_reorder(compound, u_d_mc)) %>% view()
  
  # Plot it
  ggplot(
    aes(
      x = u_d_mc,
      y = compound,
      color = terrain,
      fill = terrain
      )
  ) +
  geom_col(
    fill = "black",
    color = "black"
  ) +
  geom_errorbar(aes(
    xmin = u_d_mc - (0.5 * su_d_mc),
    xmax = u_d_mc + (0.5 * su_d_mc)),
    color = "black",
    width = 0.25
    ) +
  scale_fill_manual(values = terrain_fill_swatch) +
  scale_color_manual(values = terrain_color_swatch) +
  theme_classic() +
  labs(x = TeX("Turnover ($d^{-1}$)"),
       title = "Tundra") +
  theme(
    axis.title.y = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 1),
    axis.ticks.y = element_blank(),
    strip.background = element_blank(),
    legend.position = "none",
    plot.title = element_text(hjust = 0.5),
  )

```



## F2: Abundance versus turnover

```{r}
hull_data <- 
  turnover %>% 
  drop_na() %>% 
  group_by(terrain) %>% 
  slice(chull(compound_ug_per_g_soil, u_d_mc)) %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  ))

p_abundance_v_turnover <- turnover %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>% 
  filter(inc_time_d != 0) %>% 
  ggplot(aes(
    x = compound_ug_per_g_soil,
    y = u_d_mc,
    color = terrain
  )) +
  geom_point(
    #aes(size = dF_mc*10000),
    alpha = 0.6,
    ) +
  geom_polygon(data = hull_data,
               aes(fill = terrain,
                   color = terrain),
               show.legend = FALSE,
               alpha = 0.15,
               size = 0.15) +
  scale_color_manual(values = terrain_color_swatch) +
  theme_bw() +
  labs(x = "µg analyte (per g soil)",
       y = TeX("Turnover rate (day$^{-1}$)"),
       size = TeX("$ \\Delta ^2H$ (ppm)"),
       color = " "
  ) +
  #guides(color = guide_colorbar(barwidth = 0.5)) +
  theme(
    strip.background = element_blank(),
    legend.position = c(0.9, 0.87),
    panel.grid = element_blank(),
  )
p_abundance_v_turnover


# cowplot::save_plot(filename = "fig_output/abundance_v_turnover_1p.pdf",
#                    plot = p_abundance_v_turnover,
#                    base_height = 6,
#                    base_width = 8)
```

### Draft 2
```{r}
# Compute hull data
hull_data <- 
  turnover %>% 
  drop_na() %>% 
  group_by(terrain) %>% 
  slice(chull(compound_ug_per_g_soil, u_d_mc)) %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  ))

# Determine fit data
turnover_cleaned <- turnover %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>%
  filter(inc_time_d != 0, compound != "H2")
  
# Calculate summary statistics and export them
turnover_fit_tidy <- broom::tidy(lm(data = turnover_cleaned, u_d_mc  ~ compound_ug_per_g_soil))
turnover_fit_glance <- broom::glance(lm(data = turnover_cleaned, u_d_mc  ~ compound_ug_per_g_soil))

turnover_corr_test <- broom::tidy(
  cor.test(
    x = turnover_cleaned$compound_ug_per_g_soil,
    y = turnover_cleaned$u_d_mc,
    method = "pearson"
  )
  )

writexl::write_xlsx(turnover_fit_tidy, path = "data_output/turnover_fit_lm.xlsx")
writexl::write_xlsx(turnover_fit_glance, path = "data_output/turnover_fit_lm_stats.xlsx")
writexl::write_xlsx(turnover_corr_test, path = "data_output/turnover_corr_test.xlsx")
```

```{r}

# Plot it with the linear model!
p_abundance_v_turnover2 <- turnover %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>% 
  filter(inc_time_d != 0) %>% 
  ggplot(aes(
    x = compound_ug_per_g_soil,
    y = u_d_mc
  )) +
  # stat_density_2d(
  #   geom = "polygon", 
  #   aes(
  #     alpha = after_stat(level),
  #     group = terrain,
  #     fill = terrain
  #     ),
  #   bins = 10
  #   ) +
  # Individual data points
  geom_point(
    aes(color = terrain),
    alpha = 0.6,
    ) +
  # Using Hull Data:
  geom_polygon(
    data = hull_data,
    aes(
       fill = terrain,
       color = terrain
       ),
     show.legend = FALSE,
     alpha = 0.15,
     size = 0.15
    ) +
  geom_smooth(
    method = "lm",
    formula = "y~x",
    se = FALSE,
    color = "gray",
    linewidth = 0.3,
    linetype = "dotted"
  ) +
  scale_color_manual(values = terrain_color_swatch) +
  #scale_fill_manual(values = terrain_color_swatch) +
  #scale_x_log10() +
  #scale_y_log10() +
  #scale_alpha_continuous(range = c(0,0.5)) +
  theme_bw() +
  labs(
    x = TeX("PLFA Concentration (µg $\\cdot g^{-1}$ soil)"),
    y = TeX("Compound-specific growth rate (day$^{-1}$)"),
    size = TeX("$ \\Delta ^2H$ (ppm)"),
    color = " "
  ) +
  theme(
    strip.background = element_blank(),
    legend.position = c(0.9, 0.55),
    legend.background = element_blank(),
    panel.grid = element_blank(),
  )
p_abundance_v_turnover2

# 
# cowplot::save_plot(filename = "fig_output/abundance_v_turnover_2p.pdf",
#                    plot = p_abundance_v_turnover2,
#                    base_height = 6,
#                    base_width = 8)

```

### Inset: µ vs. 1/B

```{r}
total_biomass <- turnover_cleaned %>% 
  group_by(terrain) %>% 
  summarize(
    total_ug_per_g_soil = sum(compound_ug_per_g_soil, na.rm = TRUE)
  )

turnover_weighted_w_totals <- turnover_weighted %>% 
  filter(inc_time_d == 7) %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>% 
  left_join(total_biomass, by = "terrain")


b_ud_fit <- broom::tidy(
  lm(
    data = turnover_weighted_w_totals, 
    u_d_mc_weighted  ~ total_ug_per_g_soil
    )
  )
b_ud_stats <- broom::glance(
  lm(
    data = turnover_weighted_w_totals, 
    u_d_mc_weighted  ~ total_ug_per_g_soil
    )
)

p_turnover_totals <- turnover_weighted_w_totals %>% 
  ggplot(
    aes(
      x = total_ug_per_g_soil,
      y = u_d_mc_weighted
    )
  ) +
  geom_point(aes(color = terrain),
             size = 5) +
  # geom_smooth(
  #   method = "lm",
  #   formula = "y~x",
  #   se = FALSE,
  #   linetype = "dotted",
  #   color = "gray",
  #   size = 0.5
  # ) +
  scale_color_manual(values = terrain_color_swatch) +
  coord_cartesian(xlim = c(0, 3000), ylim = c(0.01, 0.05)) +
  labs(
    x = TeX("B (µg $\\cdot g^{-1}$ soil)"),
    y = TeX("Weighted mean \n growth rate (day$^{-1}$)"),
    color = ""
  ) +
  theme_bw() +
  theme(
    legend.position = "NA",
    axis.title = element_text(size = 10),
    axis.text = element_text(size = 10),
    axis.ticks.length = unit(-0.1, "cm"),
    panel.grid = element_blank()
  )
p_turnover_totals
```

```{r}
saveRDS(turnover_weighted_w_totals, file = "cache/turnover_weighted_w_totals.RDS")
```


### Draft 2 with inset plot!

```{r}
p_abundance_v_turnover_w_inset <- ggdraw(p_abundance_v_turnover2) +
  draw_plot(
    p_turnover_totals, 
    x = 0.65, 
    y = 0.63, 
    width = 0.33, 
    height = 0.35
    )
p_abundance_v_turnover_w_inset

cowplot::save_plot(filename = "fig_output/abundance_v_turnover_w_inset.pdf",
                   plot = p_abundance_v_turnover_w_inset,
                   base_height = 6,
                   base_width = 8)
```

### Inset: Zoom Plot (Deprecated)

```{r}
p_abundance_v_turnover_inset <- turnover %>% 
  mutate(
    terrain = case_when(
      terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
      terrain == "Marshall Mesa Grassland" ~ "Grassland",
      terrain == "Niwot Ridge Tundra" ~ "Tundra"
  )) %>% 
  filter(inc_time_d != 0) %>% 
  ggplot(aes(
    x = compound_ug_per_g_soil,
    y = u_d_mc
  )) +
  # Individual data points
  geom_point(
    aes(color = terrain),
    alpha = 0.6,
    size = 2
    ) +
  coord_cartesian(xlim = c(0, 5), ylim = c(0, 0.01)) +
  scale_color_manual(values = terrain_color_swatch) +
  theme_bw() +
  labs() +
  theme(
    strip.background = element_blank(),
    legend.position = "NA",
    legend.background = element_blank(),
    panel.grid = element_blank(),
    axis.title = element_blank(),
    panel.background = element_blank()
  )
p_abundance_v_turnover_inset
```

## F3: Literature Rates of Cell Turnover by Method

```{r lit rates}
literature_values_swatch <- c("#4ebad5", "#88e3cf", "#e388d7")
literature_values_swatch_2 <- c("#26AFE1", "#6269EB", "#E0187B")

p_lit_turnover_w_ours <- lit_turnover_w_ours %>%
  filter(method_type != "Microscopy") %>% 
  # order by mean turnover rate
  ggplot(
    aes(x = -n,
        y = u_d_mn,
        fill = method_type,
        label = round(gen_d, 1)
        )
    ) +
  geom_col() +
  #geom_text(nudge_y = 0.05, size = 2) +
  coord_flip(ylim = c(0, 0.5)) +
  scale_fill_manual(values = literature_values_swatch_2) +
  labs(y = TeX("Turnover (d$^{-1}$)"),
       fill = "Method") +
  theme_classic() +
  theme(axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 0, hjust = 0.5, size = 14),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.x = element_text(size = 18),
        #legend.position = "bottom",
        plot.title = element_text(hjust = 0.5),
        legend.position = c(0.85, 0.1),
        legend.text = element_text(size = 15),
        legend.title = element_blank(),
        plot.margin = margin(1,1,1.5,1.2, "in"))
p_lit_turnover_w_ours

# cowplot::save_plot(filename = "fig_output/literature_values.pdf",
#                    plot = p_lit_turnover_w_ours,
#                    base_height = 8,
#                    base_width = 10)
```


### Draft 2

```{r}

# Colorblind friendly palette:
lit_values_swatch_3 <- c(
  "#6269EB", # blue
  "#20968d", # gray green
  "#26AFE1", # light blue 
  "#E0187B" # pink
  )

p_lit_turnover_w_ours2 <- lit_turnover_w_ours %>% 
  mutate(
    method_type = fct_relevel(method_type, c("LH-SIP", "DNA-SIP", "TdR/Leu", "Nutrient Budgeting"))) %>% 
  arrange(method_type, desc(u_d_mn)) %>% 
  mutate(
    m = seq(1,nrow(lit_turnover_w_ours))
  ) %>% 
  ggplot(
  aes(
    x = -m,
    y = u_d_mn,
    fill = method_type,
    label = round(gen_d, 1)
    )
    ) +
  geom_col() +
  geom_text(aes(
    label = format(round(gen_d, 1)),
    color = method_type
    ),
    nudge_y = .01
    )+
  scale_fill_manual(values = lit_values_swatch_3) +
  scale_color_manual(values = lit_values_swatch_3) +
  # scale_fill_viridis_d(
  #   begin = 0,
  #   end = 0.8
  # ) +
  coord_flip(ylim = c(0, 0.4)) +
  labs(
    y = TeX("Apparent Growth Rate (day$^{-1}$)"),
    fill = "Method"
  ) +
  theme_classic() +
  theme(
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.line.y = element_blank(),
    legend.position = c(0.85, 0.2),
    legend.background = element_rect(color = "black"),
    legend.text = element_text(size = 15),
    legend.title = element_blank(),
    panel.grid.major.x = element_line(linetype = "dotted")
  ) +
  guides(
    color = "none"
  )
p_lit_turnover_w_ours2

cowplot::save_plot(filename = "fig_output/literature_values_draft2.pdf",
                   plot = p_lit_turnover_w_ours2,
                   base_height = 8,
                   base_width = 10)

# Plot annotations
lit_turnover_w_ours %>% 
  group_by(method_type) %>% 
  summarize(
    gen_d_max = max(gen_d),
    gen_d_min = min(gen_d),
    gen_d_avg = mean(gen_d)
  )

```


## SF: BacDive FA profiles


```{r}
p_tax_FA <- bacdive_FA_wide %>% 
  select(-ID) %>% 
  pivot_longer(cols = -phylum, 
               names_to = "FA",
               values_to = "rel_abund") %>%
  ggplot(aes(x = FA, 
             y = rel_abund)) +
  geom_jitter(aes(color = phylum),
              size = 0.25,
              alpha = 0.3) +
  stat_pointinterval(point_interval = "median_qi",
                     alpha = 0.65) +
  facet_wrap(vars(phylum), nrow = 3) +
  scale_color_manual(values = c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#A65628", "#FF7F00", "#F781BF", "#999999")) +
  labs(y = "Relative Abundance (%)",
       x = "",
       color = "Phylum") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90,
                                   hjust = 1),
        panel.grid = element_blank(),
        strip.background = element_blank(),
        legend.position = "NA",
        strip.text = element_text(size = 16)
        )
p_tax_FA
 
# cowplot::save_plot(filename = "fig_output/p_tax_FA.pdf",
#           plot = p_tax_FA,
#           base_height = 11,
#           base_width = 9
#           )

ggsave(p_tax_FA,
       filename = "fig_output/p_tax_FA.png",
       height = 11,
       width = 9)
```



## Compare extraction efficiency

```{r extr_eff}
FID_results_abundance_reported_long <- FID_results_abundance_no_stds %>% 
  select(
    sample_id, 
    compound, 
    deriv_efficiency, 
    extr_efficiency
  ) %>% 
  pivot_longer(
    !c(sample_id, compound),
    names_to = "stage", 
    values_to = "efficiency"
  ) %>% 
  mutate(
    stage = case_when(
      str_detect(stage, "deriv") ~ "Derivatization Efficiency",
      str_detect(stage, "extr") ~ "Extraction Efficiency"
    )
  )

FID_results_abundance_reported_long %>% 
  ggplot(aes(x = stage, y = efficiency)) +
  geom_boxplot(
    outlier.shape = NA,
    width = 0.3
    ) +
  coord_cartesian(ylim = c(0, 1.1)) +
  labs(x = "",
       y = "Efficiency") +
  theme_classic()
```



## Example chromatogram
```{r}
# plot an example chromatogram

sample_files[10] %>% # choosing arbitrary file to plot
  iso_plot_continuous_flow_data(
    # select data and aesthetics
    data = c(2), color = id1, panel = id1,
    # zoom in on time interval
    time_interval = c(750, 4000),
    # peak labels for all peaks > 2V
    peak_bounds = TRUE,
    peak_marker = FALSE,
    #peak_label_size = 3
    ) +
    scale_color_npg() +
    theme_classic() +
    theme(strip.background = element_blank())

```


## Turnover distributions


```{r turnover_across_soil}
p_turnover_distributions <- turnover %>% 
  filter(inc_time_d == 7) %>%
  mutate(terrain = case_when(
    terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
    terrain == "Marshall Mesa Grassland" ~ "Grassland",
    terrain == "Niwot Ridge Tundra" ~ "Tundra" )) %>% 
  mutate(terrain = forcats::fct_reorder(terrain, -u_d)) %>% 
  ggplot(
  aes(x = u_d, 
      y = terrain,
      color = terrain,
      fill = terrain)) +
  stat_slab(height = 0.5, 
            alpha = 0.7,
            color = "black") +
  stat_pointinterval(color = "black",
                     height = 0.2) +
  geom_point(
    color = "black",
    shape = 124,
    size = 5,
    alpha = 0.5,
    position = position_nudge(y = -.15)) +
  scale_x_log10(
    breaks = trans_breaks("log10", function(x) 10^x),
    labels = trans_format("log10", math_format(10^.x))
  ) +
  annotation_logticks(sides = "b") +
  scale_color_manual(values = terrain_color_swatch) +
  scale_fill_manual(values = terrain_fill_swatch) +
  theme_classic() +
  labs(y = "Soil Type",
       x = TeX("Turnover rate ($d^{-1}$)")) +
  theme(axis.text.y = element_text(hjust = 1),
        legend.position = "none")
p_turnover_distributions
```

## SF: Memory effect on gen_d and u_d

```{r mean_generation_time}
p_gen_time_mean <- turnover_weighted %>% 
  pivot_longer(
    cols = c(gen_d_weighted, gen_d_mc_weighted),
    names_to = "Correction",
    values_to = "gen_d"
  ) %>% 
  mutate(terrain = case_when(
    terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
    terrain == "Marshall Mesa Grassland" ~ "Grassland",
    terrain == "Niwot Ridge Tundra" ~ "Tundra" )) %>% 
  # group_by(terrain, inc_time_d) %>% 
  # summarize(terrain = terrain,
  #           gen_d_weighted = gen_d_weighted) %>% 
  # unique() %>% 
  # Plot it
  ggplot(aes(x = terrain,
             y = gen_d,
             fill = Correction)) +
  geom_col(width = 0.5,
           position = "dodge") +
  scale_y_continuous(expand=c(0,0)) +
  scale_fill_manual(values = c("#4690b0", "#3d3d3d")) +
  labs(x = "Soil Type",
       y = "Weighted Mean Generation Time (d)") +
  theme_classic() +
  theme(
    legend.position = "NA",
    axis.text = element_text(size = 6),
    axis.title = element_text(size = 8)
    )
p_gen_time_mean


p_ud_mean <- turnover_weighted %>% 
  pivot_longer(
    cols = c(u_d_weighted, u_d_mc_weighted),
    names_to = "Correction",
    values_to = "u_d"
  ) %>% 
  mutate(terrain = case_when(
    terrain == "Gordon Gulch Conifer Forest" ~ "Conifer Forest",
    terrain == "Marshall Mesa Grassland" ~ "Grassland",
    terrain == "Niwot Ridge Tundra" ~ "Tundra" )) %>%
  # group_by(terrain, inc_time_d) %>% 
  # summarize(terrain = terrain,
  #           gen_d_weighted = gen_d_weighted) %>% 
  # unique() %>% 
  # Plot it
  ggplot(aes(x = terrain,
             y = u_d,
             fill = Correction)) +
  geom_col(width = 0.5,
           position = "dodge") +
  scale_y_continuous(expand=c(0,0)) +
  scale_fill_manual(values = c("#4690b0", "#3d3d3d")) +
  labs(
    x = "Soil Type",
    y = TeX("Weighted Mean Growth Rate $(d^{-1})$")
    ) +
  theme_classic() +
  theme(
    legend.position = "NA",
    axis.text = element_text(size = 6),
    axis.title = element_text(size = 8)
    )
p_ud_mean

p_memcorr_effect <- cowplot::plot_grid(
  p_gen_time_mean,
  p_ud_mean,
  nrow = 1
)
p_memcorr_effect

cowplot::save_plot(
  plot = p_memcorr_effect,
  filename = "fig_output/memory_correction.pdf",
  base_height = 2.5,
  base_width = 6
)
```

## SF: Detection limit


```{r}
dl_df_abbreviated <- dl_df %>% 
  select(extraction_deriv_efficiency_frac, 
         detection_limit_cells,
         type) %>% 
  pivot_wider(names_from = type,
              values_from = detection_limit_cells)

p_dl <- dl_df_abbreviated %>% 
  ggplot(
    aes(x = extraction_deriv_efficiency_frac*100,
        y = average,
        color = as.factor(extraction_deriv_efficiency_frac))) +
  #geom_point(size = 1) +
  geom_errorbar(aes(
    ymin = lower, 
    ymax = upper),
    width = 0,
    size = 4) +
  # geom_hline(yintercept = 1e10,
  #            linetype = "dotted",
  #            color = "red") +
  # geom_hline(yintercept = 1e9,
  #            linetype = "dotted",
  #            color = "blue") +
  scale_x_continuous(breaks = c(20, 30, 40, 50, 60, 70, 80, 90, 100)) +
  scale_y_log10(breaks = trans_breaks("log10", function(x) 10^x),
              labels = trans_format("log10", math_format(10^.x))) +
  coord_cartesian(xlim = c(10, 110), ylim = c(1e4, 1e10)) +
  scale_color_grey(start = 0.4, end = 0.1) +
  annotation_logticks(sides = "l") +
  labs(
    x = "Combined extraction and derivatization efficiency (%)",
    y = TeX("$^2 H$ detection limit (cells)")
  ) +
  theme_bw() +
  theme(
    legend.position = "NA",
    panel.grid.minor = element_blank(),
    panel.grid.major.x = element_blank(),
    axis.ticks = element_blank(),
    axis.title = element_text(size = 10),
    axis.text.x = element_text(size = 8)
  )
p_dl

cowplot::save_plot(
  plot = p_dl,
  filename = "fig_output/detection_limit_cells.pdf",
  base_height = 4,
  base_width = 4
)

```

## SF: Fatty Acid Abundances

```{r}
p_FA <- turnover_summarized %>% 
  rename_FAs() %>% 
  filter(compound != "H2") %>% 
  filter(compound != "i-16:0") %>% 
  filter(compound != "16:1:ω5c") %>% 
  mutate(compound_edit = str_remove(compound, "-")) %>% 
  mutate(C_num = parse_number(compound_edit)) %>% 
  mutate(C_num = as.numeric(C_num)) %>% 
  mutate(compound = forcats::fct_reorder(compound, C_num)) %>% 
  ggplot(aes(x = compound,
             y = area_FID,
             fill = terrain,
             color = terrain)) +
  geom_col() +
  facet_wrap(vars(terrain, inc_time_d),
             scales = "free_y") +
  scale_fill_manual(values = terrain_fill_swatch) +
  scale_color_manual(values = terrain_color_swatch) +
  theme_classic() +
  labs(x = "Analyte",
       y = TeX("Peak Area ($pA \\cdot s$)"),
       fill = "Terrain",
       color = "Terrain") +
  theme(
    axis.text.x = element_text(angle = 90),
    legend.position = "NA"
  )
p_FA

ggsave(plot = p_FA,
       filename = "fig_output/FA_profiles.pdf",
       device = cairo_pdf,
       height = 7,
       width = 7)

# ggsave(plot = p_FA,
#        filename = "fig_output/FA_profiles.png",
#        height = 7,
#        width = 7)
```


## SF: Visible colony

```{r}
# Function Definitions
col_volumer <- function(a,b,c) {
  # Volume of a half-ellipsoid of axes a,b, and c
  vol = (4/3)*pi*a*b*c
  vol = vol/2
  return(vol)
}

cell_volumer_rod <- function(a,b,c) {
  vol = a*b*c
  return(vol)
}

cell_volumer_cocci <- function(r) {
  vol = (4/3)*pi*(r^3)
  return(vol)
}

# Define parameters of colony needed to be visible:
xax <-  1000 # x axis, y axis, z axis, in µm
yax <-  1000
zax <-  (1/6)*1000

colony_vol = col_volumer(xax, yax, zax) # generate the volume of a half-ellipsoid, circular base, 1/6th as high as wide

colony_data <- read_csv(file.path('data/bac_data.csv'))
colony_data <- colony_data %>%
  # Calculate cell volumes
  mutate(cell_volume = 
           case_when(shape == 'cocci' ~ cell_volumer_cocci(mjr_axis/2),
                     shape == 'rod' ~ cell_volumer_rod(mjr_axis, mnr_axis, mnr_axis)
                     )
         ) %>%
  # Calculate number of cells required for given volume
  mutate(num_cells_reqd = colony_vol/cell_volume) %>%
  
  # Calculate number of divisions required to reach visible colony
  mutate(divis_reqd = log2(num_cells_reqd)) %>%
  
  # Calculate hours required to reach that many divisions
  mutate(time_reqd_hr = divis_reqd*gen_time_hr)

# Get mean value for cells required in single colony
cells_in_single_colony <- colony_data %>% 
  pull(num_cells_reqd) %>% 
  mean()

colony_data_2 <- 
  tibble(gen_time_hr = seq(1, 5000, by = 1),
         num_cells_required = cells_in_single_colony,
         divis_reqd = log2(num_cells_required),
         time_reqd_hr = divis_reqd*gen_time_hr,
         gen_time_d = gen_time_hr / 24,
         time_reqd_d = time_reqd_hr / 24)
```

```{r}
p_time_reqd_for_visible_colony <- colony_data_2 %>% 
  ggplot(aes(x = gen_time_d,
             y = time_reqd_d)) +
  geom_line() +
  geom_vline(xintercept = 44.9,
             color = tundra_color,
             linetype = "dotted") +
    geom_vline(xintercept = 19.3,
             color = grassland_color,
             linetype = "dotted") +
    geom_vline(xintercept = 14.1,
             color = conifer_color,
             linetype = "dotted") +
  annotation_logticks(sides = "l") +
  scale_y_log10(breaks = trans_breaks("log10", function(x) 10^x)
              ) +
  theme_classic() +
  labs(
    x = "Maximum potential generation time (days)",
    y = "Time required for visible colony (days)")
p_time_reqd_for_visible_colony



cowplot::save_plot(
  filename = "fig_output/time_for_visible_colony.pdf",
  plot = p_time_reqd_for_visible_colony,
  base_height = 4,
  base_width = 4.5
)
```

## SF: Assimilation efficiency histogram

```{r}
# Find Fatty acid analytes manually

fas_in_d2H_assim <- d2H_assim %>% 
  filter(str_detect(Analyte, "acid") |
           str_detect(Analyte, "ate")) %>% 
  pull(Analyte) %>% 
  unique()

d2H_assim_lm <- d2H_assim %>% 
  # Select columns of interest
  select(
    org_id,
    exp_id,
    Analyte,
    `Water dD`,
    `Lipid dD`
  ) %>% 
  # Filter to only look at FAs
  filter(Analyte %in% fas_in_d2H_assim) %>% 
  drop_na() %>% 
  # Group by organism, experiment, analyte
  group_by(
    org_id,
    exp_id,
    Analyte
  ) %>% 
  # Filter out samples with three or fewer observations
  # (so that we can do a linear model)
  filter(n() > 3) %>% 
  tidyr::nest(
    data = c(`Water dD`, `Lipid dD`)
  ) %>% 
  mutate(
    fit = purrr::map(data, ~lm(`Lipid dD` ~ `Water dD`, data = .x)),
    estimates = purrr::map(fit, broom::tidy),
    summary = purrr::map(fit, broom::glance)
  ) %>% tidyr::unnest(estimates) %>%
  filter(term == "`Water dD`") %>% 
  filter(estimate >= 0,
         estimate <= 1) %>% 
  select(c(term, estimate))

mean_assim <- d2H_assim_lm %>% pull(estimate) %>% mean(na.rm = TRUE)
se_assim <- d2H_assim_lm %>% pull(estimate) %>% sd(na.rm = TRUE)

```

```{r}
p_d2H_assim_lm <- d2H_assim_lm %>% 
  ggplot(aes(x = estimate)) +
  geom_histogram(binwidth = 0.1,
                 color = tundra_color,
                 fill = tundra_color) +
  geom_vline(xintercept = mean_assim,
             linetype = "dotted") +
  geom_errorbar(
    aes(
      xmin = mean_assim - (0.5*se_assim),
      xmax = mean_assim + (0.5*se_assim),
      y = 16
        )) +
  scale_x_continuous(breaks = seq(0, 1, by = 0.1)) +
  coord_cartesian(ylim = c(0, 26),
                  expand = FALSE) +
  labs(x = "Assimilation Efficiency",
       y = "Count") +
  theme_classic()
p_d2H_assim_lm

# cowplot::save_plot(
#   plot = p_d2H_assim_lm,
#   filename = "fig_output/assim_efficiency.pdf",
#   base_height = 4,
#   base_width = 4
# )
```

## SF: IRMS Error Assessment

```{r}
# Pull the IRMS standard error for this experiment from the error of measuring standards
IRMS_standard_error <- sF_0 %>% select(F_0_at2H_se) %>% pull()
paste0("The standard error of IRMS is +/- ", IRMS_standard_error, "% (", IRMS_standard_error*10000, " ppm)")


test_standard_error_u <- tibble(
  a = 1,
  FT = IRMS_standard_error, # standard error in atom%
  F0 = 0.00000000000000000001, # start at zero
  t = seq(1, 100, by = 1), # time ranging between 0 and 100 days
  FL = 100, # label strength in atom%
  u = calculate_turnover(
    a = a,
    FT = FT,
    F0 = F0,
    t = t,
    FL = FL
  )
)

test_standard_error_u %>% 
  ggplot(
    aes(
      x = t,
      y = u
    )
  ) +
    geom_line() +
  labs(
    x = "Incubation Time",
    y = "Growth rate (day-1)",
    caption = paste0("Growth estimate resulting from", round(IRMS_standard_error*10000,4), " ppm F2H error across incubation times" )
  )
```

```{r}
error_testing <- turnover %>% 
  select(inc_time_d, dF, f_label, a, u_d_mc) %>% 
  mutate(F0 = 0) %>% 
  filter(
    inc_time_d != 0,
    !is.na(dF)) %>% 
  mutate(FT_se = IRMS_standard_error) %>% 
  mutate(
    u_d_FT_se = calculate_turnover(
      a = a, 
      FT = FT_se,
      F0 = F0,
      FL = f_label,
      t = inc_time_d
      ),
    turnover_ratio = u_d_FT_se / u_d_mc
  )


error_testing %>% 
  ggplot(
    aes(
      x = u_d_FT_se,
      fill = as.factor(inc_time_d),
      #shape = as.factor(f_label)
      )
  ) +
  geom_histogram() +
  labs(
    shape = "Label Strength (at%)",
    fill = "Incubation Time (Days)"
  )

error_testing_summary <- error_testing %>% 
  group_by(inc_time_d, f_label) %>% 
  summarize(
    u_d_FT_se_mn = mean(u_d_FT_se, na.rm = TRUE),
    FT_se_mn = mean(FT_se, na.rm = TRUE)
    ) %>% 
  rename(
    sensitivity_FT = FT_se_mn,
    sensitivity_u_d = u_d_FT_se_mn
    ) %>% 
  mutate(
    sensitivity_u_hr = sensitivity_u_d * 24,
    sensitivity_u_min = sensitivity_u_hr * 60,
    sensitivity_u_sec = sensitivity_u_min * 60
  ) %>% 
  select(inc_time_d, f_label, sensitivity_FT, sensitivity_u_d, sensitivity_u_hr, sensitivity_u_min, sensitivity_u_sec)

```

The resolution of LH-SIP is defined as the standard error of the IRMS instrument ($F_{SE}$) propagated through the turnover time calculation:

$$
\mu = - \frac{1}{t}ln(\frac{F_{SE} - a \cdot F_L}{F_0 - a \cdot F_L})
$$

Plugging in our standard error of:
```{r}
paste0("The standard error of IRMS is +/- ", IRMS_standard_error, "% (", IRMS_standard_error*10000, " ppm)")
```

```{r}
knitr::kable(error_testing_summary)
```




## SF: Sensitivity Testing

### at2H se

```{r}
LH_SIP_data %>% 
  filter(inc_time_d != 0) %>% 
  ggplot(
    aes(
      x = calibrated_at2H_without_memory_se,
      y = calibrated_at2H_with_memory_correction_se,
      color = inc_time_d_str
    )
  ) +
  geom_point(
    alpha = 0.5
  ) +
  geom_smooth(
    method = "lm",
    formula = "y~x",
    se = FALSE,
    color = "red",
    linewidth = 0.5
  ) +
  labs(
    x = TeX("SE: Without Memory Correction ($^2$H %)"),
    y = TeX("SE: With Memory Correction ($^2$H %)")
  ) +
  theme_classic()
```

### IRMS instrument error

```{r}
instrument_error <- as.numeric(isotopia::to_abundance(
  isotopia::delta(
    `2H` = 8.9, major = "1H",
    ref = "VSMOW", notation = "permil")))
instrument_error_u_d <- calculate_turnover(
  a = 0.7,
  FL = 0.5,
  F0 = 0.00000000000000000000001,
  FT = instrument_error,
  t = 7
)


instrument_error_enriched <- as.numeric(isotopia::to_abundance(
  isotopia::delta(
    `2H` = 60, major = "1H",
    ref = "VSMOW", notation = "permil")))

instrument_error_enriched_u_d <- calculate_turnover(
  a = 0.7,
  FL = 0.5,
  F0 = 0.00000000000000000000001,
  FT = instrument_error_enriched,
  t = 7
)

library(isotopia)
tibble(
  d2H_error_permil = c(8.9, 60),
  at2H_error_percent = 100*as.numeric(
    to_abundance(delta(
    `2H` = d2H_error_permil, major = "1H",
    ref = "VSMOW", notation = "permil"))
    ),
  at2H_error_ppm = 10000*at2H_error_percent,
  growth_rate_error_d = calculate_turnover(
    a = 0.7,
    FL = 0.5,
    F0 = 0,
    FT = at2H_error_percent,
    t = 7
  )
)


instrument_error_test <- crossing(
  t = seq(1, 30, by = 0.05),
  FL = c(0.1, 0.25, 0.5, 1, 2.5, 5)
) %>% mutate(
  a = 0.7,
  F0 = 0.00000000000000000000001,
  FT = instrument_error,
  u_d = calculate_turnover(
    a = a,
    FT = FT,
    F0 = F0,
    FL = FL,
    t = t
  )
)

instrument_error_test <- crossing(
  t = seq(1, 30, by = 0.05),
  FL = c(0.1, 0.25, 0.5, 1, 2.5, 5)
) %>% mutate(
  a = 0.7,
  F0 = 0.00000000000000000000001,
  FT = instrument_error,
  u_d = calculate_turnover(
    a = a,
    FT = FT,
    F0 = F0,
    FL = FL,
    t = t
  )
)

p_IRMS_error_prop <- instrument_error_test %>% 
  ggplot(
    aes(
      x = t,
      y = u_d,
      color = as.factor(FL),
      fill = as.factor(FL)
    )
  ) +
  geom_line() +
  geom_area() +
  geom_point(
    y = instrument_error_u_d,
    x = 7,
    size = 2,
    color = "black",
    shape = 1,
    fill = NA
    ) +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  scale_y_continuous(expand=c(0,0)) +
  scale_x_continuous(expand=c(0,0)) +
  theme_classic() +
  labs(
    x = "Incubation Time (days)",
    y = TeX("Growth Rate Uncertainty (day $^{-1})"),
    color = "Tracer Strength (%)",
    fill = "Tracer Strength (%)",
  ) +
  theme(
    legend.position = c(0.75, 0.5)
  )
p_IRMS_error_prop

cowplot::save_plot(
  filename = "fig_output/p_IRMS_error_prop.pdf",
  plot = p_IRMS_error_prop,
  base_height = 5,
  base_width = 6
)
```


## SF: Turnover MC effect

```{r}
turnover_summarized %>% 
  filter(inc_time_d != 0) %>% 
  ggplot(aes(
    x = u_d,
    y = u_d_mc,
    color = as.factor(inc_time_d)
  )) +
  geom_point() +
  geom_abline(linetype = "dashed") +
  labs(
    x = TeX("Growth Rate (d$^{-1}$)"),
    y = TeX("Memory Corrected (d$^{-1}$)")
  ) +
  theme_classic()
```

```{r}
turnover_summarized %>% 
  filter(inc_time_d == 7) %>% 
  pivot_longer(
    cols = c(u_d, u_d_mc),
    names_to = "corrected",
    values_to = "u_d") %>% 
  group_by(compound) %>%
  mutate(y = paste(round(mean(rt), 0), compound)) %>% 
  ungroup() %>% 
  arrange(desc(y)) %>% 
  ggplot(
    aes(
      x = u_d,
      y = y,
      color = corrected,
      size = area2
      )
  ) +
    geom_point() +
  facet_wrap(vars(terrain))
```


## SF: sµ / µ

```{r}
calculate_mu <- function(a, F2, F1, FL, t2, t1 = 0) {
  return(- (1/(t2 - t1)) * (log((a * FL - F2)/(a*FL - F1))))
}
calculate_sigma_mu <- function(a, F2, F1, FL, t2, t1 = 0, sigma_a = 0, sigma_F2 = 0, sigma_F1 = 0, sigma_FL = 0) {
  return(
    sqrt(
      ((a*FL - F2) * sigma_F1)^2 +
        
      ((a*FL - F1) * sigma_F2)^2 +
        
      (a * (F1 - F2) * sigma_FL)^2 +
        
      (FL * (F1 - F2) * sigma_a)^2
    ) / ( 
      (t2-t1) * (a * FL - F1) * (a * FL - F2) 
    )
  )
}

# error matrix
error_matrix <-
  tribble(
     ~use_a, ~use_FL, ~use_F1, ~use_F2, ~panel,
     1,       1,      1,       1,       "all errors",
     1,       0,      0,       0,       "a error only",
     0,       1,      0,       0,       "FL error only",
     0,       0,      1,       0,       "F0 error only",
     0,       0,      0,       1,       "FT error only"
  ) %>%
  mutate(panel = as_factor(panel))

# constants
model_data <- 
  tibble(
  a = 0.7069045, # assimilation factor
  t.days = 7, # days (doesn't actually matter, all cancels)
  F1 = 0, # ppm
  # errors
  sigma_a = 0.1716424,
  sigma_FL = 92, # ppm
  sigma_F1 = 1.4, # ppm
  sigma_F2 = 9.4 # ppm
) %>%
  # dynamics
  crossing(
    FL = c(5000, 10000, 25000, 50000), # ppm
    F2_interval = seq(-10, 0, by = 0.1) %>% exp()
  ) %>%
  # mu calculations
  mutate(
    F2 = F1 + FL * F2_interval,
    FL_factor = paste(FL/1e6 * 100, "at %") %>% as_factor(),
    t_factor = paste(t.days, "days") %>% as_factor(),
    mu.1_day = calculate_mu(a = a, F2 = F2, F1 = F1, FL = FL, t2 = t.days),
    TD.days = log(2) / mu.1_day
  ) %>%
  # error calculations
  crossing(error_matrix) %>%
  mutate(
    sigma_mu.1_day = calculate_sigma_mu(
      a = a, F2 = F2, F1 = F1, FL = FL, t2 = t.days,
      sigma_a = use_a * sigma_a,
      sigma_F2 = use_F2 * sigma_F2, 
      sigma_F1 = use_F1 * sigma_F1, sigma_FL = use_FL * sigma_FL
    ),
    rel_error = sigma_mu.1_day / mu.1_day
  ) %>%
  filter(!is.na(mu.1_day), !is.na(rel_error))

# plot
p_model_data <- model_data %>%
  ggplot() + 
  aes(x = TD.days / t.days, y = rel_error, color = FL_factor, linetype = t_factor) + 
  geom_line() +
  scale_x_log10(
        breaks = trans_breaks("log10", function(x) 10^x),
        labels = trans_format("log10", math_format(10^.x)),
        sec.axis = dup_axis()) +
  scale_y_continuous(labels = scales::percent) + 
  coord_cartesian(ylim = c(0, 0.5), expand = FALSE) +
  labs(
    x = latex2exp::TeX("$\\frac{T_G}{\\Delta t}$"), 
    y = latex2exp::TeX("$\\frac{\\sigma_{\\mu}}{\\mu}$"), 
    color = "Tracer Strength (%)"
  ) + 
  facet_wrap(~panel) +
  scale_color_viridis_d(option = "plasma", begin = 0, end = 0.6) +
  theme_bw() +
  theme(
    legend.position = c(0.85, 0.2),
    axis.text = element_text(color = "black"),
    plot.margin = margin(1,1,1.5,1.2, "cm")
  ) +
  guides(
    linetype = "none"
  )
p_model_data

cowplot::save_plot(
  plot = p_model_data,
  filename = "fig_output/p_model_data.pdf",
  base_height = 8,
  base_width = 10
)

```


# Supplementary Dataset Export

```{r}
# SDS1: Sent from Colorado State Lab
# SDS2: Exported BacDive Data
# SDS3: Manually-curated in Excel

# SDS4:
# SDS4.1: Summarized Growth Rates
sds4_1 <- turnover_summarized %>% 
  select(terrain, inc_time_d, compound, compound_ug_per_g_soil,
         u_d_mc, su_d_mc, gen_d_mc)

# SDS4.2: Assemblage-level growth rates
sds4_2 <- turnover_weighted %>% 
  filter(inc_time_d == 7) %>% 
  select(terrain, f_label, dF_weighted_mc, 
        u_d_mc_weighted, gen_d_mc_weighted)

# SDS4.3: IRMS Data
sds4_3 <- LH_SIP_data %>% 
  select(-c(soil, inc_time_d_str, d2H_mn_zero, d2H, at2H, compound_ug_per_g_soil, 
            `2F_Me`, `2F_FAME`, `2F_FAME_mc`, `2F_alk`, `2F_alk_mc`, `F_label_mn`,
            `F_label_se`, `F_0_at2H_se`, sa, a))

# SDS4.4: Uncertainty parameters
sds4_4 <- turnover_summarized %>% 
  filter(inc_time_d == 7) %>% 
  select(terrain, compound, 
         calibrated_at2H_with_memory_correction_se, F_label_se, 
         F_0_at2H_se, sa, su_d_mc) %>% 
  rename(
    s_ft = calibrated_at2H_with_memory_correction_se,
    s_fl = F_label_se,
    s_f0 = F_0_at2H_se,
    s_a = sa,
    s_u_d_mc = su_d_mc
    )


writexl::write_xlsx(
  sds4_1, 
  path = "data_output/sds4_1.xlsx"
  )

writexl::write_xlsx(
  sds4_2, 
  path = "data_output/sds4_2.xlsx"
  )

writexl::write_xlsx(
  sds4_3,
  path = "data_output/sds4_3.xlsx"
)

writexl::write_xlsx(
  sds4_4,
  path = "data_output/sds4_4.xlsx"
)

writexl::write_xlsx(
  s_u_d_mc_compound,
  path = "data_output/sds4_5a.xlsx"
)

writexl::write_xlsx(
  s_u_d_mc_terrain,
  path = "data_output/sds4_5b.xlsx"
)

writexl::write_xlsx(
  s_u_d_mc_total,
  path = "data_output/sds4_5c.xlsx"
)
```
