==> 99_labeling_calculations.Rmd <==
---
title: "Labeling Calculations"
date: "`r format(Sys.Date(), '%d %b %Y')`"
output:
  html_document: 
    df_print: paged
    toc: yes
    toc_float: true
    code_folding: hide
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
# libraries
library(tidyverse)
library(lubridate)
library(latex2exp)
library(isotopia) # install.packages("devtools"); devtools::install_github("isoverse/isotopia")

# global knitting options for automatic saving of all plots as .png and .pdf
knitr::opts_chunk$set(
  dev = c("png", "pdf"),
  dev.args = list(pdf = list(encoding = "WinAnsi", useDingbats = FALSE)),
  fig.keep = "all",
  fig.path = file.path("fig_output", paste0(gsub("\\.[Rr]md", "", knitr::current_input()), "_"))
)
```

# Theory

Isotopic enrichment resulting from unbiased clonal growth in the presence of an isotopic label is described by the folowing equation:

$$
F(t) = a \cdot F_L \left(1-e^{-\mu\cdot t}\right) + F_{0} \cdot e^{-\mu \cdot t}
$$
where $F_L$ is the isotopic composition (in atom % of the rare isotope) of the isotope label, $a$ is the assimilation efficiency of the tracer (from 0 to 1), $F_0$ is the isotopic composition of native biomass, $\mu$ is specific growth rate and $t$ is time. Substituting the growth rate for the expected generation time ($µ=\ln(2)/T$), one can readily calculate the required labeling time to reach a target isotopic composition $F_t$:

$$
t = \frac{T}{\ln(2)} \cdot \ln \left(\frac{a\cdot F_L - F_0}{a\cdot F_L - F_t}\right)
$$
Likewise, one can calculate the required strength of the isotope label based on expected generation time, target isotopic composition and intended incubation length:

$$
F_L = \frac{F_t - F_{0} \cdot e^{-\mu \cdot t}}{a \left(1-e^{-\mu\cdot t}\right)}
$$

# Heavy water labeling

```{r, echo=FALSE}
# parameters
dD_nat <- -200
dD_enrich_min <- +50
dD_enrich_avg <- +500
dD_enrich_max <- +5000
F_0.percent <- delta(dD_nat, ref_ratio = get_standard("2H")) %>% to_ab() %>% get_value("percent")
F_t.percent <- delta(dD_nat + dD_enrich_avg, ref_ratio = get_standard("2H")) %>% to_ab() %>% get_value("percent")
F_min.percent <- delta(dD_nat + dD_enrich_min, ref_ratio = get_standard("2H")) %>% to_ab() %>% get_value("percent")
F_max.percent <- delta(dD_nat + dD_enrich_max, ref_ratio = get_standard("2H")) %>% to_ab() %>% get_value("percent")
a <- 0.5
```


Assuming that heterotrophs incorporate the tracer from water into lipids with about a 50% efficiency (taking into consideration the combination of water hydrogen assimilation and net isotope fractionation) and that native lipids will have an isotopic composition of about $`r dD_nat`$‰, we have $a=`r a`$ and $F_0 = `r F_0.percent`$ atom %.

For precise quantification, an ideal target labeling enriches by at least $`r dD_enrich_min`$‰, at most $`r dD_enrich_max`$‰, and on average about $`r dD_enrich_avg`$‰ above the original lipids, i.e. a target of $`r dD_nat + dD_enrich_avg`$‰, or $F_t = `r F_t.percent`$ atom %. This results in the following labeling space.

## Isotope tracer strength by incubation time

```{r "tracer_strength", fig.width = 8 , fig.height = 6, warning=FALSE, message=FALSE}
# generation times of interest
gen_times <- tribble(
  ~value, ~units,
  20, "minutes",
  90,  "minutes",
  6,  "hours",
  1,  "day",
  1,  "week",
  1,  "month",
  3,  "months", 
  1,  "year",
  3,  "years"
) %>% mutate(
  label = paste(value, units) %>% factor() %>% fct_inorder(),
  gen_time.s = map2_dbl(value, units, ~duration(.x, .y)),
  gen_time.log10 = log10(gen_time.s),
  mu.1_s = log(2) / gen_time.s
)
gen_times

# incubation times
inc_time.s <- duration(c(1, 10, 50), "hours")

# labeling space
label_strengths <- 
  # generation time space (log linear to acccount for large range in gen. times)
  tibble(
    gen_time.log10 = gen_times %>% 
      with(seq(from = min(gen_time.log10), to = max(gen_time.log10), length.out = 100))
  ) %>% 
  # all combinations with incubation time
  crossing(inc_time.s) %>% 
  # min avg max quantification limits
  crossing(
    tibble(
      F_t_type = c("average", "min", "max"),
      F_t.percent = c(F_t.percent, F_min.percent, F_max.percent)
    )
  ) %>% 
  mutate(
    gen_time.s = 10^gen_time.log10,
    mu.1_s = log(2) / gen_time.s,
    F_L.percent = (F_t.percent - F_0.percent * exp(-mu.1_s * inc_time.s)) / (a * (1 - exp(-mu.1_s * inc_time.s)))
  )

# plot
label_strengths %>% 
  # spread the different F_t_type values
  select(-F_t.percent) %>%  spread(F_t_type, F_L.percent) %>% 
  # color factor
  mutate(incubation_time = sprintf("%.0f hours", inc_time.s/3600) %>% factor() %>% fct_inorder()) %>% 
  # send to plot
  ggplot() +
  aes(gen_time.log10, color = incubation_time, fill = incubation_time) + 
  geom_hline(yintercept = c(0.5, 1.0)) +
  geom_ribbon(mapping = aes(ymin = min, ymax = max, color = NULL), alpha = 0.1) +
  geom_line(mapping = aes(y = average), size = 2) +  
  labs(title = 
         sprintf("Tracer strength to reach enrichment between\n%.0f\u2030 and %.0f\u2030 (line = %.0f\u2030)", 
                 dD_enrich_min, dD_enrich_max, dD_enrich_avg)) + 
  scale_x_continuous(
    "generation time", 
    breaks = gen_times$gen_time.log10,
    labels = function(x) gen_times$label,
    expand = c(0, 0)
  ) +
  scale_y_continuous(TeX("isotope tracer strength ($F_L$)"), labels = function(x) sprintf("%.2f at%%", x)) +
  scale_color_brewer("after\nincubation\ntime", palette = "Set1") +
  scale_fill_brewer("after\nincubation\ntime", palette = "Set1") +
  coord_cartesian(ylim = c(0, 1)) +
  #facet_wrap(~incubation_time, nrow = 1) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid.minor = element_blank()) 
```

## Incubation time by tracer strength

```{r "req_incubation_time", fig.width = 8 , fig.height = 6, warning=FALSE, message=FALSE}
# labeling strengths
F_L.percent <- c(0.5)

# incubation times
inc_times <- 
  # generation time space (log linear to acccount for large range in gen. times)
  tibble(
    gen_time.log10 = gen_times %>% 
      with(seq(from = min(gen_time.log10), to = max(gen_time.log10), length.out = 100)) %>% 
      union(gen_times$gen_time.log10)
  ) %>% 
  # all combinations with tracer strength
  crossing(F_L.percent) %>% 
  # min avg max quantification limits
  crossing(
    tibble(
      F_t_type = c("average", "min", "max"),
      F_t.percent = c(F_t.percent, F_min.percent, F_max.percent)
    )
  ) %>% 
  mutate(
    gen_time.s = 10^gen_time.log10,
    inc_time.s = gen_time.s / log(2) * log( (a * F_L.percent - F_0.percent) / (a * F_L.percent - F_t.percent) ),
    inc_time.log10 = log10(inc_time.s)
  )

# y breaks
y_breaks <- inc_times %>% 
  filter(F_t_type == "average") %>% 
  semi_join(gen_times, by = "gen_time.log10") %>% 
  mutate(
    # clean up lubridate time estimates
    label = duration(inc_time.s) %>% as.character() %>% 
      str_extract("(\\~\\d+\\.?\\d*\\s\\w+|.*s$)") %>% 
      str_replace("\\.\\d+s", " seconds")
  )

# plot
inc_times %>% 
  # spread the different F_t_type values
  select(-F_t.percent, -inc_time.s) %>%  spread(F_t_type, inc_time.log10) %>% 
  # color factor
  mutate(label_strength = paste("F_L =", F_L.percent, "at%") %>% factor() %>% fct_inorder()) %>% 
  # send to plot
  ggplot() +
  aes(gen_time.log10, color = label_strength, fill = label_strength) + 
  geom_hline(
    data = tibble(inc_time.s = as.numeric(inc_time.s), inc_time.log10 = log10(inc_time.s)),
    mapping = aes(yintercept = inc_time.log10, linetype = paste(inc_time.s/3600, "hours"))
  ) +
  geom_ribbon(mapping = aes(ymin = min, ymax = max), alpha = 0.1) +
  geom_line(mapping = aes(y = average), size = 2) +  
  labs(title =
         sprintf("Required incubation times to reach enrichment\nbetween %.0f\u2030 and %.0f\u2030 (thick line = %.0f\u2030)",
                 dD_enrich_min, dD_enrich_max, dD_enrich_avg),
       linetype = "target incubation times") +
  scale_x_continuous(
    "generation time", 
    breaks = gen_times$gen_time.log10,
    labels = function(x) gen_times$label,
    expand = c(0, 0)
  ) +
  scale_y_continuous(
    "required incubation time", 
    breaks = y_breaks$inc_time.log10,
    labels = function(x) y_breaks$label,
    expand = c(0, 0)
  ) +
  scale_color_brewer("tracer strength", palette = "Set1") +
  scale_fill_brewer("tracer strength", palette = "Set1") +
  coord_cartesian(ylim = range(y_breaks$inc_time.log10)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid.minor = element_blank())
```

```{r}
# plot
inc_times %>% 
  # spread the different F_t_type values
  select(-F_t.percent, -inc_time.s) %>%  spread(F_t_type, inc_time.log10) %>% 
  # color factor
  mutate(label_strength = paste("F_L =", F_L.percent, "at%") %>% factor() %>% fct_inorder()) %>% 
  # send to plot
  ggplot() +
  aes(
    gen_time.log10, 
    color = label_strength, 
    fill = label_strength
    ) + 
  geom_ribbon(
    mapping = aes(ymin = min, ymax = max), 
    alpha = 0.1
    ) +
  geom_line(
    mapping = aes(y = average), 
    size = 2) +  
  labs(title =
         sprintf("Required incubation times to reach enrichment\nbetween %.0f\u2030 and %.0f\u2030 (thick line = %.0f\u2030)",
                 dD_enrich_min, dD_enrich_max, dD_enrich_avg),
       linetype = "target incubation times") +
  scale_x_continuous(
    "generation time", 
    breaks = gen_times$gen_time.log10,
    labels = function(x) gen_times$label,
    expand = c(0, 0)
  ) +
  scale_y_continuous(
    "required incubation time",
    breaks = y_breaks$inc_time.log10,
    labels = function(x) y_breaks$label,
    expand = c(0, 0)
  ) +
  scale_color_brewer("tracer strength", palette = "Set1") +
  scale_fill_brewer("tracer strength", palette = "Set1") +
  coord_cartesian(ylim = range(y_breaks$inc_time.log10)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5, size = 12),
        axis.text.y = element_text(size = 12),
        axis.title = element_blank(),
        panel.grid.minor = element_blank())
```



## Enrichment curves by generation time

```{r "enrichment_curves", fig.width = 8 , fig.height = 7, warning=FALSE, message=FALSE}
# labeling strengths
F_L.percent <- c(0.5, 1)

# labeling curves
labeling_curves <- 
  # extend to twice max incubation time
  tibble(inc_time.s = seq(from = 0, to = 2 * max(inc_time.s), length.out = 100)) %>%   
  crossing(gen_times[-1, ]) %>% 
  crossing(F_L.percent) %>% 
  mutate(
    F_t.percent = a * F_L.percent * (1 - exp(-mu.1_s  * inc_time.s)) + F_0.percent * exp(-mu.1_s * inc_time.s),
    inc_time.h = inc_time.s/3600,
    grid_y = paste("F_L =", F_L.percent, "at%")
  )

# plot
p1 <- labeling_curves %>% 
  # panel for time zoom
  mutate(grid_x = ifelse(inc_time.h <= 5.1, "1", "2")) %>% 
  # make sure time transition at panel edge is smooth
  { bind_rows(., filter(., grid_x == "1") %>% filter(inc_time.h == max(inc_time.h)) %>% mutate(grid_x = "2")) } %>% 
  # send to plot
  ggplot() + 
  aes(inc_time.h, F_t.percent, color = label) +
  # highlight area of quantification
  geom_rect(xmin = -Inf, xmax = Inf, ymin = F_min.percent, ymax = F_max.percent, 
            color = "gray", fill = "gray", alpha = 0.25, show.legend = FALSE) + 
  geom_hline(yintercept = F_0.percent, linetype = 2) +
  geom_hline(yintercept = F_t.percent) +
  #geom_vline(xintercept = as.numeric(inc_time.s, "hours")) +
  geom_line() + 
  labs(title = "Isotopic composition/enrichment after incubation time\n(gray box = quantification range)") + 
  scale_x_continuous("Incubation time [hours]", expand = c(0, 0)) +
  scale_y_continuous(
    TeX("estimated isotopic composition ($F_t$)"), labels = function(x) sprintf("%.2f at%%", x),
    sec.axis = sec_axis(
      trans = function(x) { 
        ab(x, notation = "percent") %>% 
          to_ratio() %>% to_delta(ref_ratio = get_standard("2H")) %>% 
          get_value("permil") - 
        ab(F_0.percent, notation = "percent") %>% 
          to_ratio() %>% to_delta(ref_ratio = get_standard("2H")) %>% 
          get_value("permil")
      }, 
      labels = function(x) sprintf("%+.0f\u2030", x), 
      name = TeX("estimated enrichmen ($\\delta{}D$)")),
    expand = c(0, 0)
  ) +
  scale_color_brewer("generation time", palette = "Set1") +
  facet_grid(grid_y ~ grid_x, scales = "free_x") +
  theme_bw() +
  theme(strip.background.x = element_blank(), strip.text.x = element_blank(),
        panel.spacing.x = unit(0, "mm"))

p1 + geom_hline(
  mapping = aes(yintercept = 0.325, linetype = "approx upper\nlimit for IRMS")
) + scale_linetype_manual(values = 2)
```

### Zoom in on quantification range

```{r "enrichment_curves_zoomed1", fig.width = 8 , fig.height = 7, warning=FALSE, message=FALSE}
p1 + coord_cartesian(ylim = c(0, 0.12))
```

### Zoom in on low enrichment

```{r "enrichment_curves_zoomed2", fig.width = 8 , fig.height = 7, warning=FALSE, message=FALSE}
p1 + coord_cartesian(ylim = c(0.01, 0.025), xlim = c(0, 300))
```
